#274: rating component hooks
